Route protectors added

+++ client/src/services/admin/categorieApi.js
@@ -2,10 +2,19 @@
 import axios from "axios";
 
 const API_BASE_URL = "http://localhost:5193/api/v1";
+const TOKEN = localStorage.getItem("authToken");
 
 export const getCategories = async () => {
+  if (!TOKEN) {
+    throw new Error("No token found");
+  }
+
   try {
-    const response = await axios.get(`${API_BASE_URL}/categories`);
+    const response = await axios.get(`${API_BASE_URL}/categories`, {
+      headers: {
+        "Authorization": `Bearer ${TOKEN}`,
+      },
+    });
     if (response.status === 200) {
       return response.data;
     }
@@ -16,11 +25,16 @@ export const getCategories = async () => {
 };
 
 export const createCategorie = async (categoriesData) => {
+  if (!TOKEN) {
+    throw new Error("No token found");
+  }
+
   try {
     let data = JSON.stringify(categoriesData);
     const response = await axios.post(`${API_BASE_URL}/categories`, data, {
       headers: {
         "Content-Type": "application/json",
+        "Authorization": `Bearer ${TOKEN}`,
       },
     });
 
@@ -32,12 +46,17 @@ export const createCategorie = async (categoriesData) => {
 };
 
 export const editCategorie = async (codeCategorie, categoriesData) => {
+  if (!TOKEN) {
+    throw new Error("No token found");
+  }
+
   try {
     let data = JSON.stringify(categoriesData);
     const response = await axios.put(
       `${API_BASE_URL}/categories/${codeCategorie}`, data, {
         headers: {
           "Content-Type": "application/json",
+          "Authorization": `Bearer ${TOKEN}`,
         },
       });
     return response;
@@ -48,9 +67,17 @@ export const editCategorie = async (codeCategorie, categoriesData) => {
 };
 
 export const deleteCategorie = async (codeCategorie) => {
+  if (!TOKEN) {
+    throw new Error("No token found");
+  }
+
   try {
     const response = await axios.delete(
-      `${API_BASE_URL}/categories/${codeCategorie}`
+      `${API_BASE_URL}/categories/${codeCategorie}`, {
+        headers: {
+          "Authorization": `Bearer ${TOKEN}`,
+        },
+      }
     );
     return response;
   } catch (error) {
