+++ client/src/services/admin/itineraireApi.js
@@ -2,10 +2,19 @@
 import axios from "axios";
 
 const API_BASE_URL = "http://localhost:5193/api/v1";
+const TOKEN = localStorage.getItem("authToken");
 
 export const getItineraires = async () => {
+  if (!TOKEN) {
+    throw new Error("No token found");
+  }
+
   try {
-    const response = await axios.get(`${API_BASE_URL}/itineraires`);
+    const response = await axios.get(`${API_BASE_URL}/itineraires`, {
+      headers: {
+        "Authorization": `Bearer ${TOKEN}`,
+      },
+    });
     if (response.status === 200) {
       return response.data;
     }
@@ -16,11 +25,16 @@ export const getItineraires = async () => {
 };
 
 export const createItineraire = async (itinerairesData) => {
+  if (!TOKEN) {
+    throw new Error("No token found");
+  }
+  
   try {
     let data = JSON.stringify(itinerairesData);
     const response = await axios.post(`${API_BASE_URL}/itineraires`, data, {
       headers: {
         "Content-Type": "application/json",
+        "Authorization": `Bearer ${TOKEN}`,
       },
     });
 
@@ -32,14 +46,22 @@ export const createItineraire = async (itinerairesData) => {
 };
 
 export const editItineraire = async (codeItineraire, itinerairesData) => {
+  if (!TOKEN) {
+    throw new Error("No token found");
+  }
+
   try {
     let data = JSON.stringify(itinerairesData);
     const response = await axios.put(
-      `${API_BASE_URL}/itineraires/${codeItineraire}`, data, {
+      `${API_BASE_URL}/itineraires/${codeItineraire}`,
+      data,
+      {
         headers: {
           "Content-Type": "application/json",
+          "Authorization": `Bearer ${TOKEN}`,
         },
-      });
+      }
+    );
     return response;
   } catch (error) {
     console.error("Erreur lors de l'envoi des données:", error);
@@ -48,9 +70,18 @@ export const editItineraire = async (codeItineraire, itinerairesData) => {
 };
 
 export const deleteItineraire = async (codeItineraire) => {
+  if (!TOKEN) {
+    throw new Error("No token found");
+  }
+
   try {
     const response = await axios.delete(
-      `${API_BASE_URL}/itineraires/${codeItineraire}`
+      `${API_BASE_URL}/itineraires/${codeItineraire}`,
+      {
+        headers: {
+          "Authorization": `Bearer ${TOKEN}`,
+        },
+      }
     );
     return response;
   } catch (error) {
